# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mentors:
    get:
      operationId: listMentors
      tags: [mentors]
      summary: List active mentors
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MentorProfile"
    post:
      operationId: createMentor
      tags: [mentors]
      summary: Create a mentor (admin)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MentorProfileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /mentors/{id}:
    get:
      operationId: getMentor
      tags: [mentors]
      summary: Get a mentor with articles
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateMentor
      tags: [mentors]
      summary: Update a mentor (admin)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MentorProfileInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteMentor
      tags: [mentors]
      summary: Delete a mentor (admin)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Mentor Articles ────────────────────────────────────────
  /mentors/{mentorId}/articles:
    get:
      operationId: listMentorArticles
      tags: [mentor-articles]
      summary: List articles for a mentor
      parameters:
        - in: path
          name: mentorId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MentorArticle"
    post:
      operationId: createMentorArticle
      tags: [mentor-articles]
      summary: Create article for a mentor (admin)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
        - in: path
          name: mentorId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MentorArticleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorArticle"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /mentor-articles/{id}:
    put:
      operationId: updateMentorArticle
      tags: [mentor-articles]
      summary: Update an article (admin)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MentorArticleInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorArticle"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteMentorArticle
      tags: [mentor-articles]
      summary: Delete an article (admin)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Job Listings ───────────────────────────────────────────
...
  /startup-applications:
    post:
      operationId: createStartupApplication
      tags: [startup-applications]
      summary: Submit a startup application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                founderName:
                  type: string
                  minLength: 1
                email:
                  type: string
                  minLength: 1
                registrationStatus:
                  type: string
                  minLength: 1
                startupIdea:
                  type: string
                  minLength: 1
                readiness:
                  type: string
                  minLength: 1
                readinessDetail:
                  type: string
                  default: ""
                ideaReason:
                  type: string
                  minLength: 1
                experience:
                  type: string
                  minLength: 1
                team:
                  type: string
                  minLength: 1
              required:
                - founderName
                - email
                - registrationStatus
                - startupIdea
                - readiness
                - ideaReason
                - experience
                - team
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartupApplication"
    get:
      operationId: listStartupApplications
      tags: [startup-applications]
      summary: List startup applications (admin)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StartupApplication"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /startup-applications/{id}/result:
    patch:
      operationId: updateStartupApplicationResult
      tags: [startup-applications]
      summary: Set result for a startup application (admin)
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                result:
                  type: string
                  enum: [도전가능, 도전불가능]
                resultReason:
                  type: string
                  default: ""
              required: [result]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartupApplication"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /startup-applications/{id}:
    get:
      operationId: getStartupApplication
      tags: [startup-applications]
      summary: Get a startup application by id (public – for applicant result view)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartupApplicationPublic"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Creative Works ────────────────────────────────────────
...
components:
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Mentor Application ─────────────────────────────────
...
    MentorProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        major:
          type: string
        yearsOfExperience:
          type: integer
        photoUrl:
          type: ["string", "null"]
        headlineText:
          type: string
        sublineText:
          type: string
        bio:
          type: string
        avatarColor:
          type: string
        initial:
          type: string
        isActive:
          type: boolean
        displayOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - major
        - yearsOfExperience
        - photoUrl
        - headlineText
        - sublineText
        - bio
        - avatarColor
        - initial
        - isActive
        - displayOrder
        - createdAt
    MentorProfileInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        major:
          type: string
          minLength: 1
        yearsOfExperience:
          type: integer
        photoUrl:
          type: ["string", "null"]
        headlineText:
          type: string
        sublineText:
          type: string
        bio:
          type: string
        avatarColor:
          type: string
        initial:
          type: string
          minLength: 1
        isActive:
          type: boolean
        displayOrder:
          type: integer
      required: [name, major, yearsOfExperience, initial]
    MentorDetail:
      allOf:
        - $ref: "#/components/schemas/MentorProfile"
        - type: object
          properties:
            articles:
              type: array
              items:
                $ref: "#/components/schemas/MentorArticle"
          required: [articles]

    # ── Mentor Article ─────────────────────────────────────
    MentorArticle:
      type: object
      properties:
        id:
          type: integer
        mentorId:
          type: integer
        title:
          type: string
        content:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
      required: [id, mentorId, title, content, displayOrder, isActive]
    MentorArticleInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
      required: [title]

    # ── Job Listing ────────────────────────────────────────
...
    StartupApplication:
      type: object
      properties:
        id:
          type: integer
        founderName:
          type: string
        email:
          type: string
        registrationStatus:
          type: string
        startupIdea:
          type: string
        readiness:
          type: string
        readinessDetail:
          type: string
        ideaReason:
          type: string
        experience:
          type: string
        team:
          type: string
        result:
          type: ["string", "null"]
        resultReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - founderName
        - email
        - registrationStatus
        - startupIdea
        - readiness
        - readinessDetail
        - ideaReason
        - experience
        - team
        - result
        - resultReason
        - createdAt
    StartupApplicationPublic:
      type: object
      properties:
        id:
          type: integer
        founderName:
          type: string
        result:
          type: ["string", "null"]
        resultReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, founderName, result, resultReason, createdAt]

    # ── Creative Work ────────────────────────────────────────
...